therd commit

+++ app/Http/Controllers/Admin/ProductsController.php
@@ -6,9 +6,12 @@ use App\Http\Controllers\Controller;
 use App\Models\Category;
 use App\Models\Product;
 use App\Models\ProductImage;
+use App\Models\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class ProductsController extends Controller
 {
@@ -19,10 +22,16 @@ class ProductsController extends Controller
      */
     public function index()
     {
+        $this->authorize('view-any',Product::class);
+
         $products = Product::with('category')
             ->latest()
             ->orderBy('name', 'ASC')
+            //->withoutGlobalScopes()
+            //->Status('draft')
             ->paginate(5);
+
+            
         return view('admin.products.index', [
             'products' => $products,
             'categories' => Category::all(),
@@ -37,10 +46,16 @@ class ProductsController extends Controller
      */
     public function create()
     {
+        $this->authorize('create',Product::class);
+
+        /*if(!Gate::allows('product.create'))
+        {
+            abort(403);
+        }*/
         return view('admin.products.create', [
             'product' => new Product(),
             'categories' => Category::all(),
-
+            'tags' => '',
         ]);
     }
 
@@ -52,6 +67,12 @@ class ProductsController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create',Product::class);
+
+       /* if(Gate::denies('product.create'))
+        {
+            abort(403);
+        }*/
         $request->validate(Product::validateRules());
 
         $request->merge([
@@ -77,6 +98,8 @@ class ProductsController extends Controller
         /*$product=new Product($request->all());
         $product->save();*/
 
+        $product->tags()->attach($this->getTags($request));
+
         return redirect()->route('admin.products.index')
             ->with('success', "Product ($product->name) created!");
     }
@@ -90,6 +113,9 @@ class ProductsController extends Controller
     public function show($id)
     {
         $product = Product::findOrFail($id);
+
+        $this->authorize('view',$product);
+
         return view('admin.products.show', [
             'product' => $product,
         ]);
@@ -103,10 +129,17 @@ class ProductsController extends Controller
      */
     public function edit($id)
     {
+         //Gate::authorize('product.update');
+
         $product = Product::findOrFail($id);
+
+        $this->authorize('update',$product);
+
+        $tags = $product->tags()->pluck('name')->toArray();
         return view('admin.products.edit', [
             'product' => $product,
             'categories' => Category::all(),
+            'tags' => implode(',', $tags),
         ]);
     }
 
@@ -119,12 +152,17 @@ class ProductsController extends Controller
      */
     public function update(Request $request, $id)
     {
+        //Gate::authorize('product.update');
+
         $product = Product::findOrFail($id);
+
+        $this->authorize('update',$product);
+
         $request->validate(Product::validateRules());
 
         $data = $request->all();
-        
-        $previous=false;
+
+        $previous = false;
 
         if ($request->hasFile('image')) {
             $file = $request->file('image');
@@ -134,7 +172,7 @@ class ProductsController extends Controller
             //store in public disk
             //$data['image'] = $file->store('/images','public');
             //OR
-            $data['image'] = $file->store('/images',[
+            $data['image'] = $file->store('/images', [
                 'disk' => 'uploads',
             ]);
 
@@ -144,34 +182,34 @@ class ProductsController extends Controller
                 'disk' => 'uploads',
             ]);*/
 
-            $previous=$product->image;
+            $previous = $product->image;
         }
         $product->update($data);
         //OR
         //$product->fill($request->all())->save();
 
-        if($previous){
+        if ($previous) {
             Storage::disk('uploads')->delete($previous);
         }
 
         //Gallery(multiple file)
-        if($request->hasFile('gallery')){
-            foreach($request->file('gallery')as $file){
-               $image_path = $file->store('/images',[
+        if ($request->hasFile('gallery')) {
+            foreach ($request->file('gallery') as $file) {
+                $image_path = $file->store('/images', [
                     'disk' => 'uploads',
                 ]);
-               /* $product->images()->create([
+                /* $product->images()->create([
                     'image_path'=>$image_path,
                 ]);*/
                 //OR
-                $image=new ProductImage([
-                    'image_path'=>$image_path,
+                $image = new ProductImage([
+                    'image_path' => $image_path,
                 ]);
                 $product->images()->save($image);
             }
         }
-        
 
+        $product->tags()->sync($this->getTags($request));
 
         return redirect()->route('admin.products.index')
             ->with('success', "Product ($product->name) updated!");
@@ -186,14 +224,49 @@ class ProductsController extends Controller
     public function destroy($id)
     {
         $product = Product::findOrFail($id);
+        $this->authorize('delete',$product);
         $product->delete();
 
-        if($product->image)
-        {
+        if ($product->image) {
             Storage::disk('uploads')->delete($product->image);
         }
 
         return redirect()->route('admin.products.index')
             ->with('success', "Product ($product->name) deleted!");
+    }
+
+    protected function getTags(Request $request)
+    {
+        $tag_ids = [];
+        $tags = $request->post('tags');
+        $tags = json_decode($tags);
+        //DB::table('product_tag')->where('product_id','=',$product->id)->delete();
+        if (count($tags) > 0) {
+
+            foreach ($tags as $tag) {
+                $tag_name = $tag->value;
+                $tagModel = Tag::firstOrCreate([
+                    'name' => $tag_name,
+                ], [
+                    'slug' => Str::slug($tag_name),
+                ]);
+                //OR
+                /*$tagModel=Tag::where('name',$tag_name)->first();
+            if(!$tagModel){
+                $tagModel=Tag::create([
+                    'name'=>$tag_name,
+                    'slug'=>Str::slug($tag_name),
+                ]);
+            }*/
+                /*DB::table('product_tag')->insert([
+                'product_id'=>$product->id,
+                'tag_id'=>$tagModel->id,
+
+            ]);*/
+
+            $tag_ids[] = $tagModel->id;
+            }
+        }
+        return $tag_ids;
     }
 }
